+++ tiger-common/src/main/java/de/gematik/test/tiger/common/data/config/CfgTigerProxyOptions.java
@@ -5,18 +5,22 @@ import lombok.Data;
 
 @Data
 public class CfgTigerProxyOptions {
+
+    /**
+     * Management-port of the Tiger Proxy.
+     */
     private int serverPort = -1;
 
     /**
-     * used to overwrite proxyCfg with values that allow to reverse proxy the given server node
+     * used to overwrite proxyCfg with values that allow to reverse proxy the given server node.
      */
     private String proxiedServer;
 
     /**
-     * used to overwrite proxyCfg with correct route table, as we cant know whether its http or https, we introduced
-     * this field and default to http
+     * Used to add a route to the tiger-proxy. By default, the healthcheck-url-protocol is used here, or http if none is
+     * present. If you want to override this you can do it using this field.
      */
-    private String proxyProtocol = "http";
+    private String proxiedServerProtocol;
 
     TigerProxyConfiguration proxyCfg;
 }