+++ ReleaseNotes.md
@@ -5,6 +5,8 @@
 ## Breaking Change
 * TGR-113: Test-Context and Test-Variables are deprecated. All values are now stored using `TigerGlobalConfiguration`. This supersedes all uses of domains and context, of which there are no known instances. If your migration is difficult, please contact the team.
 * The migration entails a complete rethink of configuration and value-stores in tiger. If you have any troubles please read the user-manual and don't hesitate to ask us.
+* TGR-299: The configuration flag `proxyProtocol` from the server-type tigerProxy has been renamed to `proxiedServerProtocol` to clarify usage and avoid confusion.
+* If a `proxiedServer` did have a Healthcheck-URL (or source-URL for `externalUrlServer`) previously the reverse-proxy would target the deep-path. Now it will target the domain only.
 
 ## Bugfix
 * TGR-350: Multiple Scenariooutlines in feature file break test execution.